+++ packages/celest_core/example/android/app/src/main/kotlin/dev/celest/celest_core_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.celest.celest_core_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()